make ValidCharactersRule more flexible
It now allows the set of valid characters to be set
via a pattern in the config and for overrides to be
set for given tag values.

This is mostly to support an internal decision to relax
group names so that they allow `^` and `~`. Others values
should not be relaxed as it could result in backwards
compatibility issues for names that were already being
converted to `_` when published.