allow preflight check to be used separately
Moves the handler for CORS preflight checks to
CustomDirectives and adds a `cors` helper for
wrapping a route with both the preflight handler
and filter to add CORS headers to the inner route.
This change allows the CORS support to be used
without pulling in all of the standard options
from the default request handler.