support inline rollup of datapoints (#525)
This change allows datapoints to be rolled up as they
are flowing into the in memory storage. Typically this
has been done later as part of offline processing using
EMR or via a separate aggregator cluster that would
perform the aggregations prior to being sent to the
in memory storage.

To be eligible for the rollup the data must have already
been normalized to the step boundary and rate conversion
for counters should have already been performed.

The inline rollup will create a rollup block so it will
have more overhead in terms of memory use unless the
reduction multiple is over 4. Data will only go through
the rollup path if it is tagged with `atlas.rollup=true`.

The publish payload can now take an `id` field which will
get used to dedup messages. If used, then it should avoid
over counting if retries are present in the publishing
pipeline as long as the updates for a given datapoint are
always mapped to the same id for a given interval. The
message id should not be used for multiple intervals or
subsequent messages would incorrectly get lost.