add limit operation that can be used with sort
Adds a `:limit` operation that can be used with the
existing `:sort` and `:order` to restrict the output
set. The `:head` operator has been removed as a data
operator and changed to be an alias for limit.

Limit was chosen because it is more familiar to those
with a SQL background. The `:head` operator was never
really optimized in the data layer and a desire to use
sort with a limit for top-N use-cases is more prominent
than the sampling use-cases for which `:head` was
originally intended, but that did not really materialize
in practice.