enable self-reporting for standalone example
The standalone example will not report data to itself
when running with the sample `memory.conf`. A similar
thing can be done with the `lwcapi.conf` by uncommenting
the enabled flag. It is disabled by default since there
are some ordering issues on startup that cause exceptions
in the log.