named rewrite expr string now preserves offset (#540)
Fixes a bug with the expression string for named rewrites
that contain offsets. If the rewrite is being used similar
to an aggregate function and the display expression is a
query, then the offset was getting lost because that is an
attribute of the aggregate functions that are only present
in the expanded evaluation expression.

Now the named rewrite will extract the offset and group by
information if the display type is a query. For others the
rewrite will be re-evaluated using the same interpreter.