reduce allocations for processing messages (#541)
Some initial performance checks on the streaming
evaluation show a lot of allocations. The biggest
contributor was converting from ByteString to
a form that jackson could parse. This has been
reduced by reusing buffers. Some other minor
changes have been made to reduce allocations for
fields and to go direct to SmallHashMap for the
tag data.

Biggest bottlenecks right now are parsing the
json messages from the SSE stream followed by
the materialization speed. The materialization
speed should be less of an issue in practice
because the streams will be longer lived and is
supposed to be improved in akka 2.5.0.