simplify eval usage for common use-cases
Adds an `Evaluator` class with a simple public API
for evaluating the stream. It is a java class that
exposes the reactive streams types so it is easy to
use with various implementations. There are two
main methods:

**createPublisher(uri)**

Creates a publisher for a graph uri so the same data
can be accessed as a stream. This makes it really easy
for a user to transition from one view to another.

```
+------------+
|  URI       |
|  Eval      | -------- [TimeSeriesMessage] --->
|  Source    |
+------------+
```

**createStreamsProcessor**

Takes a set of id/URI pairs as inputs and evaluates
all of the expressions. This allows for better reuse
when evaluating many expressions. There is an id
for each input that will be added to the output so
that corresponding entries can be correlated.
The complete set of current data sources must be
supplied as part of each input event.

```
                       +------+
                       |      |
--- [DataSources] ---> | Eval | --- [MessageEnvelope] --->
  Set((id,URI), ...)   |      |   (id, TimeSeriesMessage)
                       +------+
```