fix multi-threaded reads from long set/map (#603)
Several use-cases of the primitive collections assume
multi-threaded read-only access is safe after the collection
is built and effectively immutable. That is nothing will
ever write to it again. The variants with a Long key type
use an internal buffer to avoid allocations when passing
the long value to `MurmurHash3.bytesHash`. The multi-threaded
access could result in errors such as BufferOverflowException
on the buffer.

This changes those collections to use a ThreadLocal for keeping
track of the buffers.