change ids to custom class
Before the ids were just a BigInteger. This was mostly
for convenience and it was an easy starting point for
some modifications.

Flame graphs of stateful clusters for some stacks are
showing a significant about of time being spent computing
the hash code for the ids when looking up the blocks. The
custom class in this change uses a precomputed hash code
so that it will be constant time.