fix regex matching when prefix is an exact match
When searching for the prefix it was looking for the postion
of the first match that was strictly greater than the prefix.
For cases where the prefix is an exact match to the full value
they would not match. The search now checks for the first
match that is greater than or equal to the prefix.