add :clamp-min and :clamp-max
Allows the user to restrict the min and max values of
the input respectively. The most common use-case motivating
this is to be able to support a bounded auto-scaling for
data on an axis. The axis lower and upper limits are either
explicit or automatic. These operators give more flexiblity
and can be set for a given line to tune its behavior.

Name was chosen to match the same operators in Promoetheus.