fix error when rendering with a long time range
When trying to render a chart with a long time range
a NoSuchElementException would result if the graph
width was too small. Sample exception below:

```
java.util.NoSuchElementException: head of empty list
	at scala.collection.immutable.Nil$.head(List.scala:428) ~[scala-library-2.12.2.jar:?]
	at scala.collection.immutable.Nil$.head(List.scala:425) ~[scala-library-2.12.2.jar:?]
	at com.netflix.atlas.chart.graphics.Ticks$.time(Ticks.scala:265) ~[atlas-chart_2.12-1.6.0-SNAPSHOT.jar:1.6.0-SNAPSHOT]
```

The cause is that there were no options for meeting
the desired number of major ticks given the options
for timeTickSizes. This change uses a simpler selection
method with just major ticks for longer ranges setting
it to the name of the month or year.