add mappings for application load balancer stats (#634)
Adds basic mappings for the ApplicationELB metrics in
CloudWatch. There was a bit of debate on whether the
LoadBalancer and TargetGroup dimensions should be kept
as is or extract the name and only show that. The ids
are rarely useful to the end user and add quite a bit
of noise when selecting in the UI. For now though we
just map the values in as they are provided in CloudWatch.