add helper for simple shard mapping
Move over the internal helper utility for mapping data
to a set of instances. This is typically used with a
service like edda that maintains stable slotting for
instances within auto-scaling groups. These simple
schemes are preferred for the core monitoring system
to avoid dependencies on complex infrastructure like
zookeeper that are prone to having correlated failure
with the key systems we need to monitor.