fix endless loop for numProbesPerKey (#652)
This method was incorrectly using the data length rather
than the capacity when looping over the array. If a key
had a collision and was hashed to a position that was
larger than the data length, then it would be an endless
loop because the key entry would never be found.