scalafmt config to help ensure consistency
This is being used on some of the internal projects
across the team. Should help with consistency, but
makes it harder to investigate the git history because
of all the changes for the initial format. To avoid
that problem in the future the format is checked as
part of the PR build.