reduce thread contention in SubscriptionManager
Refactors the subscription manager to make the updates
cheaper and minimize the synchronization that is needed.
Now the only synchronization is to ensure the that the
cleanup for the sub handlers map will not have a race
condition with adding new handlers.

This change also avoids the more costly computation of
the subscription handlers from the registration map. On
a test node with real traffic it significantly reduced
the CPU and thread contention observed.