custom map wrapper to improve get performance
For some use-cases when feeding into streaming evaluator
there are a lot of get calls on the underlying tag maps
for the datapoints. This change adds a custom map wrapper
that maps the `get(key)` call on the java map to the
`getOrNull(key)` on the SmallHashMap. This avoids the
allocation of the Option when using the default scala
wrapper.