send exceptions from db back to sender
This should fix #684. Before the exception was getting
ignored leaving the user to wait for a timeout and not
providing any context.

As a side note this was already resolved in some of the
internal wrappers, this change makes it send back a
Failure type on exception to stay compatible with the
internal step and allow us to keep reducing the differences.