fix possible buffer overflow for refreshing groups (#704)
The EurekaGroupsLookup stage creates a source to regularly
refresh the set of Eureka vips that are needed based on the
requested data sources. When a new set of data sources comes
in, it will stop the old source and create a new one.

The problem is when a new set of data sources comes in quickly
it could stop the source refreshing Eureka groups after the
request was made and before the response was processed. In
this scenario the `max-open-requests` limit would eventually
be reached and all subsequent requests would fail.

There are two changes included here:

1. Before the source was stopped using a `takeWhile` operator
   at a later stage. Now it is stopped by changing the source
   to indicate it is done and no longer has data. This means
   that the subsequent stages of the flow should go ahead and
   run and so the response body of any outstanding requests will
   get processed. The BufferOverflowException is still possible
   with just this change if too many changes occur in a short
   period, however, they will eventually clear up if the rate
   goes down.
2. The input is now throttled and conflated so that if many
   changes to the set of data sources come in a short window
   all but the last will get ignored.