fix stop for Eureka refresh source (#726)
In #704, there was a change to fix how the Eureka refresh
source was stopped. That change has a bug in that the function
passed to `repeatWhile` has a reference to the member variable.
So it is almost always true, unless it happens to get checked
in the brief period between when the previous instance was set
to false and the new instance is created.

As a result, none of the sources are actually stopped and the
old DataSources instances get pushed out repeated. Depending
on how the expressions have changed, this can cause the
subscriptions for various expression to come and go.

This change avoids the problem on the actual instance. Still
need to come up with a way to test, but will do that as a new
PR so I can unblock some downstream users that need the fix.