reduce thread contention for splitter cache (#732)
On instance type with a large number of cores the cache
introduced in #730 starts to make things worse because
of heavy thread contention with the `computeIfAbsent`
calls used for interning the queries and internally to
the LoadingCache. This change refactors the calls to
do a get/put which may result in recomputing some values,
but avoids the blocking. See Netflix/spectator#352 for
a similar issue we had before.

This change was manually pushed to a test node and I
confirmed that the blocked threads are no longer present.