fix #733, add :freeze operator
Adds `:freeze` operator to allow freezing a portion
of the stack so it is isolated from further modification.
This is useful for query overlay and some other use-cases
where we essentially need to combine multiple expressions
and still be able to use operations like `:clear` or `:list`
without worrying about exactly how far back to go on the
stack.