add mapping for TrustedAdvisor metrics
For now it is only the service limit usage metrics that
give us an idea of how close we are to limits. This can
be a simple alternative to many of the uses of howlermonkey.
The update frequency for the trusted advisor metrics is
fairly low so it will take an hour or so to detect a
problem. For slow growth it is adequate.