add helper function for simplifying a query
A number of internal libraries have the use-case to
perform rewrites and then simplify the query expression.
This change adds it as a common helper function to the
Query object. In the case of rewrites it is often useful
to ignore terms so it has the option to use special
handling of the not true case so it will get elided rather
than making the overall expression false.