limit size of in clause for expansion (#771)
When building the query index we expand `:in` clauses
so we get a set of simpler queries with `:eq` clauses.
This allows us to greatly reduce the number of queries
that wind up in the fallback set that have to be checked
linearly.

However, if there are many `:in` clauses and some of them
match a large set, then there is a combinatorial explosion
and the number of queries we need to index becomes too
large. This changes restricts the expansion to `:in` clauses
with 5 or less values.