fix #761, allow basic math ops on StyleExpr
This change relaxes the language restrictions a bit
so that basic math operations can can be applied to
a time series expression after style attributes have
been applied. The main use-case is to allow for some
UI operations that need to perform operations on an
arbitrary graph in a similar manner to using `:cq`
to restrict the scope.

For binary operations, only one side can be a StyleExpr.
This prevents mutually exclusive presentation attributes
from the two sides.