fix :cq with custom averages (#777)
If a `:cq` operation was used to restrict a query
by a tag that was not available on the denominator,
then it would result in a no data line because the
query would get applied to both sides. This change
simplifies the rewrite so it will always re-run
the operation after modifying the display expression.
In this case that means the `extractCommonQuery`
step will get executed on the new query formed by
applying the `:cq` operation.