add group by keys to time series message
Fixes #754. Exposes the set of keys that were used
explicitly as part of a group by in the time series
messages emitted by fetch and streaming evaluation.

This is useful for things like canary analysis that
generate a set of lines for multiple clusters and
need a reliable way to find the correlated entries
from the baseline and canary. Since the tags will
contain exact matches from the query as well as
keys used in the group by, this change will allow
the user to extract just the keys from the group by
and ignore other exact match tags coming from the
scope used on the query.