indicate DST transition on time axis
A frequent question after a DST transition is why the
time axis labels are off. This change indicates the DST
transition by changing the first major tick label after
to show `+1H` or `-1H`. The color is red to make it
standout from other tick lables on the same axis.