add helper for tracking index stats
This provides a quick summary of some key stats for
the index to see what is contributing the most to the
number of distinct strings. In most cases we just use
hive queries for digging into this, but these stats
can be used before data is in hive and are always scoped
to the specific index. In some cases where we already
have rollup policies the hive data is raw from before
the additional policies are enabled.