set dstype for polled cloudwatch metrics (#827)
It was not set before so all would get treated as rate.
This led to some confusion for some of the S3 metrics
where it appeared there was a massive drop in the amount
of data stored in s3 after a brief gap. This drop was
because the consolidation averaged with 0 for the NaN
values since the values were considered rates.