honor legend flag for v2.json format image (#830)
See #564 and #568 for more context. The included image would
suppress the image. For some use-cases it is desirable to
include it. The previous output can still be achieved by
explicitly using `no_legend=1`.

Main use-case right now is an explode view that shows the
images, but has some dynamic filtering for outliers or missing
data. This avoids the need to make multiple requests for the
same data if the user requests to see the legends.