fix state isolation for streaming execution (#831)
If an expression has multiple equivalent stateful
sub-expressions, then the they would store the state
to the same place since the key is the stateful expr
instance. This changes the map used to look at the
identity when performing the state lookup, so separate
instances will have isolated state.

We'll probably need to refactor this more thoroughly
long term and either memoize equivalent sub-expressions
during the evaluation or lookup based on position.