update travis to build on jdk10 (#833)
Dropping jdk9 as there is no reason to use it over
jdk10 at this point.

One test case needed to be fixed because there
is now a `java.version.date` system property and
it was accessing `java.version` expecting a string
instead of an object.

The tests fail on OpenJDK because of font diffs.
For more information see:

http://mail.openjdk.java.net/pipermail/discuss/2017-November/004621.html

Equals verifier is not yet compatible with jdk11
bytecode. Also some image diffs are present when
using the early access OracleJDK 11. So for now
The build will only run on OracleJDK 10.