config setting for window of time grouping
Adds a config setting that can be used to set the number
of buffers that will be maintained when grouping the
datapoints into time buckets. The default is 1 which will
flush the data as soon as data arrives for the next time
frame. More buffers allows more time for data to accumulate,
but means that it will take longer to emit data for a given
time.