short circuit pattern matching with prefix check (#858)
In the streaming execution, up to 25% of the CPU usage is
checking regular expressions. This change improves the
chances that checks will be able to quickly fail without
needing regex by always checking the fixed prefix when
known in the matches function. This wasn't done before
because when evaluating with the index we already know
the prefix matches before checking.