make random operator deterministic (#860)
The `:random` operator is mostly used for generating sample
data that has some noise. However, in its current form the
values will change every time it runs. When experimenting with
expressions this makes it hard to reason about because the
input is always changing.

Now it will be based on a hash of the timestamp so it is
reproducible. There is also now a seeded variant, `:srandom`
that can be used if multiple sample lines are needed with
different noise.

Fixes #847.