add step parameter to `DataSource` (#865)
As part of enabling high resolution metrics for the streaming path, this
commit adds a step parameter to `DataSource`. This is then used to
configure the step in the metrics subscription.

I tested this by running an lwcapi server locally and observing that the
step listed in the response from `/api/v1/expressions` matched the
overridden value.

Note that this is an incremental step. For this to fully function, the
step needs to be introduced into the `Flow` that groups and
evaluates the datapoints. This will be done in a follow up commit.