re-register handler if needed on heartbeat (#891)
There is a race condition on reconnects where the new
connection can come in before the stream for the old
connection is closed. The cleanup for the old connection
will then unregister the stream id. This is a short-term
work around to ensure running streams have a registered
handler and track the number of events that we see.