remove support for post to StreamApi (#892)
Simplify the stream api and reduce duplication with subscribe.
The only use right now is the basic get with subscriptions
managed via the subscribe api. The plan is to switch to a
websocket via subscribe that will accept subscription updates
and send back the stream of results. For now just cleaning
up some unused aspects of the current setup.