fix `LoadBalancer` extractor pattern
Configure the app|net matching as a non-capturing group, so the correct
portion is extracted.

As I was thinking about the addition of NLB metrics, I realized the
previous commits don't handle conditional mapping of `LoadBalancer`
which will need to set `aws.alb` or `aws.nlb` depending on the context.
In the course of writing a passing case for `aws.alb` and a failing one
for `aws.nlb`, this bug was uncovered. I'll fix the ALB vs NLB handling
in a separate commit.