add directives for setting Netflix-Endpoint header
To improve the quality of metrics it is useful to be able
to see what path is being used. However, it shouldn't match
the full path that might have ids or other elements that
would make the cardinality explode. This change adds path
directives that will capture specified static portions and
map those to the `Netflix-Endpoint` header.

Note, when nested with other usages of the standard path
directives it will capture extracted portions from those
directives and include it in the header. While not ideal,
it isn't a common case and since the IPC logger has limiting
it will mostly impact the user with the broad matching.