add hook to control startup delay for lwcapi
This delay is used to keep the service in an unhealthy
state for a specified window. Clients that publish data
use the load balancer and will not start publishing until
it is healthy. The eval library that does the subscriptions
will attemtp to connect once it detects the new instance.
This delay means that subscriptions should be in place
before data starts flowing.