aggregate datapoints while grouping
This helps reduce the amount of memory required when
evaluating an expression. If a 1k node cluster is
sending data for a single sum expression, then it will
now result in a single value being maintained in the
time grouping stage rather than keeping one per node
until the final evaluation.