add helpers for rolling min/max
Also sets up a base OnlineAlgorithm trait that can be
used to get better reuse consuming the operations. The
state can be captured as a Config object to make it easy
to save the previous state and restore if needed.

Will refactor DES helpers to use the same pattern in a
follow up PR.