add helper to compute a delayed signal (#912)
Delays the values by the window size. This is similar to
the `:offset` operator except that it can be applied to
any input line instead of just changing the time window
fetched with a DataExpr. Short delays can be useful for
alerting to detect changes in slightly shifted trend lines.

This change adds the online algorithm implementation, the
actual StatefulExpr will be done as a follow up.