Add optional CloudWatch metric timeout (#915)
This commit adds a mechanism to specify how long the system should
interpolate a base value for unreported CloudWatch metrics before
ceasing to send them.

CloudWatch will return 0 metrics in at least two cases:
  1. No metrics were recorded.
  2. The resource has been removed, but metrics for it still fall within
     the retention window.

I've made this optional and configurable at the metric category level to
tactically address the issue for RDS reported by a user and to minimize
visual clutter in the config. Depending on usage in practice, we may
want to require it to be specified for every category or lift the scope
to make it consistent at the global level.