fix behavior for aggregating count (#917)
The aggregate datapoint will already have been converted
from a raw value to a base count at the source. So when
it is received here it should behave just like a simple
sum. Before it was converting the datapoint to a 1 or 0
when aggregating as well which can lead to drastically
smaller count than expected if the source has many time
series for the query.