add ignore and pipeline helpers for step alignment
Adds an online algorithm to ignore the first N values. This
can be used with sliding DES to align to a step boundary
before starting the DES computation.

Also adds a pipeline that can be used to chain together a
sequence of online algorithms while conforming to the same
interface.