Report `NaN` until first datapoint
Report `NaN` until first datapoint for metric categories with a timeout
configured.

I reordered the tests to more closely follow the actual flow of data
from startup and the progression to no data for a deleted resource.

I've also noted the rationale for interpolating `0` values for gaps in
CloudWatch data.