initial websocket support for lwcapi (#924)
Adds a websocket endpoint `/api/v1/subscribe` that can be
used for both updating the set of subscriptions and receiving
the stream of data. This also begins some changes to help
improve consistency with other Atlas APIs, namely the messages
rename `frequency` to `step` and the `/lwc` prefix is removed
from the endpoint path. For now there is an alias so the eval
client will still work with the messages coming from older
lwcapi services running.

This change also use the model objects from the eval client
on the server side for the websocket endpoint to help ensure
better consistency of the messages in the future. The old
endpoints have not been changed to preserve compatibility
during the transition.

Current plan:

1) Deploy updated lwcapi service with the new endpoint. The
   client will continue to work as before.
2) Update the eval client and get internal usages migrated.
3) Cleanup the lwcapi to remove the old subscribe/stream
   endpoints.

This should be the final big change before releasing 1.6.