diagnostic messages from publisher to subscriber (#925)
Adds support for diagnostic messages to be sent to the
evaluate api along with the data values. This can be used
for the publisher to send diagnostic messages that will
be received by a particular subscriber. An example use-case
is to reject an expression due to the amount of load already
on the publisher.