lwc: server to client time notifications (#927)
The eval client for consuming LWC data uses the timestamps
in the messages so it can be used to run on live or captured
data that may have old timestamps. If there is a subscription
to an expression that doesn't match any data, then no messages
will go through and thus the time grouping will not be flushed
for a given interval.

This change allows the server to send heartbeats with the
timestamp and step so it can ensure there will always be at
least one message coming though for a given stream. When
running on previously captured data it will also include the
heartbeat messages and exhibit the same behavior.