do not request gzip for /stream call
Current testing indicates the largest chunk of CPU time
server side is in the GZIP compression. The akka directive
doesn't easily allow for configuring it with `BEST_SPEED`
instead of `BEST_COMPRESSION`. Further on the client side,
the largest chunk of allocations is for the decompression.

Overall it seems preferable right now to use more bandwidth
and avoid the compression. Will revisit after web-socket
change and switching to binary encoding for stream.