switch to source based on BlockingQueue (#939)
The default `Source.queue` has quite a bit of overhead and
the input must be throttled based on the future to bound the
memory use. `Source.actorRef` has a bit less overhead, but
has the same unbounded memory use if it cannot keep up and
does not provide a reasonable way to throttle the input.

This change uses a custom source stage based on the java
ArrayBlockingQueue. There is significantly less overhead and
the memory usage is bounded even with an offer and forget
usage style.