improve performance of findValues for common tags (#949)
For tags that are repeated across many metrics, this change
prunes the item set by doing a lookup for that key value. The
`andNot` is quite a bit faster than iterating the matches.
It mostly impacts queries that are projecting all values for
a common tag without other restrictions.

For the sample benchmark, findValuesAllOne is the only one
to show a noticeable change as expected. Before it was about
4864.472 ops/s and after 357671.851 ops/s.

Trying it on real data, it is mostly faster, but there are a
handful of cases where it is a bit slower. The table below
shows mean time per operation in milliseconds for a handful
of common queries:

| *URI*                | *Before*  | *After*   |
|----------------------|-----------|-----------|
| `/nf.app`            | 115.5     | 26.9      |
| `/nf.cluster`        | 102.0     | 66.3      |
| `/nf.node`           | 156.4     | 196.9     |
| `/nf.account`        | 128.6     | 2.0       |
| `/nf.region`         | 97.1      | 2.3       |
| `/nf.zone`           | 90.5      | 5.5       |
| `/name`              | 122.5     | 111.0     |

Full URI is with prefix of `/api/v1/tags`.