only use error images for browsers (#955)
There are some programmatic use-cases for accessing images
such as alerting emails and bots for including in chat apps.
The current behavior makes it hard to detect errors with the
images because it is embedded within the image text and has
a 200 response code. This changes the graph api behavior to
only use error images when the access appears to be coming
from a web browser.