fix #948, all zeros shown on y-axis
If the upper bound exactly matched `10 * factor` for the
selected unit prefix, then it would use a different selection
to avoid large numbers for the tick labels. However, for an
exact match it is better to use the default prefix to avoid
getting zeros due to rounding with the larger prefix.