build using openjdk11 (#973)
This updates the travis builds to use OpenJDK 11. The
`-release 8` option is used to ensure the generated
bytecode will still work on JDK8. Due to font rendering
differences, image tests will fail when running on older
versions of the JDK or on operating systems other than
Mac OS X. Those checks will now automatically be
disabled on systems that are known to fail, but are checked
as part of CI validation.