avoid conversion if already a ConfigValue
In the docs site it is getting config 1.2 in the sbt
classpath. There isn't an obvious way to force it to
a newer version. The older version will fail if trying
to convert a ConfigValue to a ConfigValue. For now we
can workaround the problem by special casing that to
avoid the unnecessary conversion.