use dedicated object for algo state (#978)
Before it was using a Config object for convenience. This
switches it to a dedicated object that can be easily used
with `Json.encode/decode` or other similar tools. This should
also be more efficient for the more common use-cases because
we can avoid creation of the needless config objects.