Throttle calls to CloudWatch
We're hitting CloudWatch rate limits on a regular basis. However, the
AWS limits should be sufficient for our overall per second call rate in
the majority of cases. The current pattern of calls has bursts when the
`Tick` message kicks off a collection, which causes the call rate to
spike above the per second limit. This commit introduces call rate
limiting to smooth out the request pattern.

The Akka documentation for throttling request/response actor
communication is incomplete. Through iteration playing around in a local
toy app, I arrived at the implementation herein and confirmed that it
works as expected.

To further improve resource utilization, I plan to migrate to the batch
API for metric data in a later commit.