Throttle calls to CloudWatch (#983)
We're hitting CloudWatch rate limits on a regular basis. However, the
AWS limits should be sufficient for our overall per second call rate in
the majority of cases. The current pattern of calls has bursts when the
`Tick` message kicks off a collection, which causes the call rate to
spike above the per second limit. This commit introduces call rate
limiting to smooth out the request pattern.

The Akka documentation for throttling request/response actor
communication is incomplete. Through iteration playing around in a local
toy app, I arrived at the implementation herein and confirmed that it
works as expected.

For this use case, it's important to ensure that either all or none of
the `MetricMetadata` elements are added for processing. To satisfy that
requirement, the full list is sent to the actor `Source` which then uses
`flatMapConcat` to send each element individually through the throttle
phase. This provides a stronger guarantee than the default, which could
drop elements if the queue fills up. In practice, memory is more likely
to be the limiting factor, given actors have unbounded mailboxes by
default. Case in point, it was difficult to trigger the drop scenario in
the local toy app. However, this approach more deterministically
provides the stronger guarantee.