consistent state model for all stateful operators
Updates all of the stateful operators to use the same
online algorithm base classes. This also gives them a
consistent representation of state that can easily be
serialized and deserialized. This is a first step to
possible future work of persisting the state of streaming
evaluations so it can be replayed or the execution can
be transitioned to another instance.