Improve CloudWatch metric lag handling
This is the first of potentially multiple commits to improve handling
of CloudWatch metric lag. This first commit makes the time range of
the query configurable and adds a metric to track the age (in
CloudWatch periods) of the latest datapoint returned. This will enable
assessing the distribution of ages across the namespaces and metrics
collected. Those data will influence the approach for datapoints that
are older than Atlas will accept.