Fix lambda config
The code that loads `MetricCategory` instances from the config expects
the `dimension` field, even if it's empty. This was causing the service
to silently fail.

In addition to fixing the `lambda` config, I've added tests that load
the production config to improve the chances we'll catch this at build
time. I've also added logging, so the problem will be evident in the
logs.