Fix lambda config (#1010)
The code that loads `MetricCategory` instances from the config expects
the `dimension` field, even if it's empty. This was causing the service
to silently fail.

In addition to fixing the `lambda` config, I've added tests that load
the production config to improve the chances we'll catch this at build
time.