improve handling of empty legend string
Before it would fail with an error that is hard for a user
to understand:

```
IllegalArgumentException: Can't add attribute to 0-length text
```

This error would only show up if the image legend was
rendered causing further confusion because it wouldn't
show up if the legend was suppressed.

Now it will fallback to the default legend string if an
empty string is specified.